Organize plugins paths in a reusable way
