+++ C2GUILauncher/src/Chivalry2Launchers.cs
@@ -20,6 +20,6 @@ namespace C2GUILauncher
         /// <summary>
         /// The modded launcher is used to launch the game with mods. The DLLs here are the relative paths to the DLLs that are to be injected. Just a stub right now.
         /// </summary>
-        public static ProcessLauncher ModdedLauncher { get; } = new ProcessLauncher(GameBinPath, BinDir, new string[] { "XAPO.dll" });
+        public static ProcessLauncher ModdedLauncher { get; } = new ProcessLauncher(GameBinPath, BinDir, new string[] { Mods.CoreMods.InjectorDllPath });
     }
 }
