+++ C2GUILauncher/src/MainWindow.xaml.cs
@@ -13,6 +13,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using C2GUILauncher.Mods;
 
 namespace C2GUILauncher
 {
